enh(secu): sanitize and cast parameters to int
Convert the num, limit, and period parameters to int. Sanitize them and call
intval on them, when performing the XML export on the event logs page.

These parameters are certainly meant to be numeric, since we do calculations
with them later on, such as this:

    $limitReq = " LIMIT " . $num * $limit . ", " . $limit;

Casting them to int prevents the possibility for SQL injection in the line
above. Additionally, on line 656 we loop over values derived from $num, and
that can lead to an infinite loop if $num is not numeric.

This change causes no functional changes to the event logs XML export. When
calling this page with non-numeric values for num, limit, or period, the
default options are used instead.